refactor(web/applications): move create-new-case folder to root (boas-740)
