+++ apps/web/src/server/applications/case/documentation/applications-documentation.router.js
@@ -1,6 +1,6 @@
 import { Router as createRouter } from 'express';
 import asyncRoute from '../../../lib/async-route.js';
-import { assertDomainTypeIsNotInspector } from '../../pages/create-new-case/applications-create.guards.js';
+import { assertDomainTypeIsNotInspector } from '../../create-new-case/applications-create.guards.js';
 import * as locals from '../applications-case.locals.js';
 import applicationsDocumentationMetadataRouter from '../documentation-metadata/documentation-metadata.router.js';
 import * as controller from './applications-documentation.controller.js';
