+++ apps/web/src/server/applications/case/edit/applications-edit.router.js
@@ -1,6 +1,6 @@
 import { Router as createRouter } from 'express';
-import * as guards from '../../pages/create-new-case/applications-create.guards.js';
-import { registerCaseId } from '../../pages/create-new-case/applications-create.locals.js';
+import * as guards from '../../create-new-case/applications-create.guards.js';
+import { registerCaseId } from '../../create-new-case/applications-create.locals.js';
 import applicationsEditApplicantRouter from './applicant/applications-edit-applicant.router.js';
 import { registerBackPath } from './applications-edit.locals.js';
 import applicationsEditCaseRouter from './case/applications-edit-case.router.js';
