+++ apps/web/src/server/applications/create-new-case/applicant/applications-create-applicant.guards.js
@@ -1,8 +1,8 @@
-import pino from '../../../../lib/logger.js';
+import pino from '../../../lib/logger.js';
 import {
 	getSessionApplicantInfoTypes,
 	getSessionCaseHasNeverBeenResumed
-} from '../../../common/services/session.service.js';
+} from '../../common/services/session.service.js';
 import * as applicationsCreateApplicantService from './applications-create-applicant.service.js';
 
 /**
