+++ apps/web/src/server/applications/create-new-case/key-dates/applications-create-key-dates.router.js
@@ -1,6 +1,6 @@
 import { Router as createRouter } from 'express';
-import asyncRoute from '../../../../lib/async-route.js';
-import { registerCaseWithQuery } from '../../../applications.locals.js';
+import asyncRoute from '../../../lib/async-route.js';
+import { registerCaseWithQuery } from '../../applications.locals.js';
 import * as locals from '../applicant/applications-create-applicant.locals.js';
 import * as controller from './applications-create-key-dates.controller.js';
 import * as validators from './applications-create-key-dates.validators.js';
