+++ apps/web/src/server/applications/create-new-case/key-dates/applications-create-key-dates.types.d.ts
@@ -1,4 +1,4 @@
-import { CaseCreateProps } from '../../../applications.types';
+import { CaseCreateProps } from '../../applications.types';
 
 export interface ApplicationsCreateKeyDatesBody extends Record<string, string | undefined> {
 	'keyDates.submissionDatePublished'?: string;
